+++ packages/pigment-css-react/src/Stack.d.ts
@@ -1,0 +1,17 @@
+import * as CSS from 'csstype';
+
+import { Breakpoint } from './base';
+import { PolymorphicComponent } from './Box';
+
+type CssProperty<T> = T | Array<T> | Partial<Record<Breakpoint, T>>;
+
+type StackBaseProps = {
+  spacing?: CssProperty<number | string>;
+  direction?: CssProperty<CSS.StandardLonghandProperties['flexDirection']>;
+  divider?: React.ReactNode;
+  className?: string;
+};
+
+declare const Stack: PolymorphicComponent<StackBaseProps>;
+
+export default Stack;