+++ packages/pigment-css-react/src/index.ts
@@ -1,3 +1,4 @@
+export * from './base';
 export { default as styled, type StyledComponent } from './styled';
 export { default as sx } from './sx';
 export { default as keyframes } from './keyframes';
@@ -5,5 +6,5 @@ export { generateAtomics, atomics } from './generateAtomics';
 export { default as css } from './css';
 export { default as createUseThemeProps } from './createUseThemeProps';
 export { default as internal_createExtendSxProp } from './createExtendSxProp';
-export { default as Box } from './Box';
 export { default as useTheme } from './useTheme';
+export { default as globalCss } from './globalCss';