+++ packages/pigment-css-react/src/utils/convertAtomicsToCss.ts
@@ -7,12 +7,15 @@ export type Atomics = {
     [key: string]: string[];
   };
   shorthands: Record<string, string[]>;
+  multiplier?: string;
 };
 
 export type RuntimeConfig = {
   conditions: string[];
   styles: Record<string, Record<string, Record<string, string>>>;
   shorthands: Atomics['shorthands'];
+  defaultCondition: string;
+  multiplier?: string;
 };
 
 function getClassName(...items: string[]) {
@@ -20,7 +23,13 @@ function getClassName(...items: string[]) {
 }
 
 export function convertAtomicsToCss(
-  { conditions = {}, defaultCondition, properties, shorthands = {} }: Atomics,
+  {
+    conditions = {},
+    defaultCondition,
+    properties,
+    shorthands = {},
+    multiplier = undefined,
+  }: Atomics,
   mainClassName: string,
   isGlobal = false,
   debug = false,
@@ -30,6 +39,8 @@ export function convertAtomicsToCss(
     styles: {},
     shorthands,
     conditions: Object.keys(conditions),
+    defaultCondition,
+    multiplier,
   };
   let count = 1;
   function getCount() {
@@ -46,6 +57,11 @@ export function convertAtomicsToCss(
   Object.entries(conditions).forEach(([conditionName, mediaQueryStr]) => {
     Object.entries(properties).forEach(([cssPropertyName, propertyValues]) => {
       propertyValues.forEach((propertyValue) => {
+        const propValue = propertyValue.startsWith('--')
+          ? cssesc(
+              `var(${propertyValue}${conditionName === defaultCondition ? '' : `-${conditionName}`})`,
+            )
+          : propertyValue;
         const className =
           isGlobal || debug
             ? getClassName(
@@ -60,7 +76,7 @@ export function convertAtomicsToCss(
           classes.push({
             className,
             css: {
-              [cssPropertyName]: propertyValue,
+              [cssPropertyName]: propValue,
             },
           });
         } else {
@@ -68,19 +84,19 @@ export function convertAtomicsToCss(
             className,
             css: {
               [mediaQueryStr]: {
-                [cssPropertyName]: propertyValue,
+                [cssPropertyName]: propValue,
               },
             },
           });
         }
-        const classMap = runtimeConfig.styles[cssPropertyName] ?? {};
-        const conditionClassMap = classMap[propertyValue] ?? {};
-        conditionClassMap[conditionName] = className;
-        if (conditionName === defaultCondition) {
-          conditionClassMap.$$default = className;
+
+        if (!runtimeConfig.styles[cssPropertyName]) {
+          runtimeConfig.styles[cssPropertyName] = {};
         }
-        classMap[propertyValue] = conditionClassMap;
-        runtimeConfig.styles[cssPropertyName] = classMap;
+        if (!runtimeConfig.styles[cssPropertyName][propertyValue]) {
+          runtimeConfig.styles[cssPropertyName][propertyValue] = {};
+        }
+        runtimeConfig.styles[cssPropertyName][propertyValue][conditionName] = className;
       });
     });
   });