+++ packages/pigment-css-react/tsup.config.ts
@@ -20,7 +20,15 @@ const baseConfig: Options = {
   external,
 };
 
-const BASE_FILES = ['index.ts', 'theme.ts', 'Box.jsx', 'RtlProvider.tsx'];
+const BASE_FILES = [
+  'index.ts',
+  'theme.ts',
+  'Box.jsx',
+  'RtlProvider.tsx',
+  'Stack.jsx',
+  'Container.jsx',
+  'Hidden.jsx',
+];
 
 export default defineConfig([
   {